+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for "Hello Matrix" Bot
 
+## Changes currently in development
+
+- Kanban Tool integration now supports following boards for changes
+
+
 ## Changes in v0.1 (2016-11-09)
 
 - Initial release